Rename this package from `LSFClusterManagers.jl` (plural) to `LSFClusterManager.jl` (singular)
